v0.83.0: test coverage, API polish & code quality
Implements all 25 roadmap items for v0.83.0:

Test Coverage (FUZZ-BULK-01, PROPTEST-02, CDC-ASYNC-01, KFAIL-DOC-01, REG-TESTS-01, ERRPATH-01, DATALOG-MAXITER-TEST-01):
- Add fuzz targets for N-Triples, N-Quads, TriG bulk loaders
- Add oxigraph reference-implementation proptest (ntriples_oxigraph.rs)
- Add CDC LISTEN/NOTIFY barrier integration test (no sleep)
- Annotate known_failures.txt entries with # Reason: / # Issue: comments
- Add pg_regress tests for 13 previously untested pg_extern functions
- Add 8 error-path regression tests
- Add seminaive max-iteration guard regression test

API (API-RENAME-01, API-GRAPH-COL-01):
- Add load_jsonld() SQL alias; json_ld_load() emits NOTICE deprecation
- Confirm find_triples() g column present (no change needed)

Code Quality (MOD-BIDI-01, GUC-NAME-01, CHANGELOG-BREAK-01, CHANGELOG-FMT-01, DEPAUDIT-01, RENOVATE-01, P-05-EVAL):
- Split src/bidi.rs into 5 sub-modules (protocol/relay/subscribe/sync/mod)
- Document GUC naming convention in CONTRIBUTING.md
- Adopt BREAKING- Adopt BREAKING- Adopt BREAKING- Adopted
- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- Ad- vate.json for automated dependency updates
- Write plans/p05_shar- Write plans/p05_shar- Write plans/p05_shar- Write plans/p05_shar-, MER- Write plans/p05_shar- Write plans/p05_shar- Write plao_div- Write plans/p05_shar- Write plans/p05_shar- Write plans/p05_shar- Woff_secs GUC

Security / HTTP (BUILD-TIME-FIELD-01, HTTP-401-WWW-AUTH-01, Security / HTTP (BUILD-TIME-FIEDOC-0Security / HTTP (BUILD-TIME-FIELD-01, HTTP-401-WWWeld is now RFC 3339 timestamp
- WWW-Authen- WWW-Authen- WWW-Authen- WWW-Authen- WWW-Authen- WWW-Authen- WWWenvelo- WWW-Authen- WWW-Authen- WWW-Authen- WWW-Authen- WWW-Aut
- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl- Bl-HANGELOG.md: v0.83.0 entry
- ROADMAP.md: v0.83.0 marked Released
