+++ pg_ripple_http/src/routing/admin_handlers.rs
@@ -15,7 +15,10 @@ use super::ApiDoc;
 use utoipa::OpenApi as _;
 
 /// Build timestamp recorded at compile time (RFC 3339).
-const BUILD_TIME: &str = env!("CARGO_PKG_VERSION"); // fallback; real build time requires build.rs
+/// BUILD-TIME-FIELD-01 (v0.83.0): populated by build.rs with an RFC-3339
+/// timestamp, or falls back to the Cargo package version string prefixed with
+/// "build-version=" when SOURCE_DATE_EPOCH is unset.
+const BUILD_TIME: &str = option_env!("BUILD_TIMESTAMP").unwrap_or(concat!("build-version=", env!("CARGO_PKG_VERSION")));
 
 pub(crate) async fn health(State(state): State<Arc<AppState>>) -> Response {
     // v0.55.0 I-3: return structured JSON with version, git_sha, postgres_connected, last_query_ts.
