+++ tests/w3c/known_failures.txt
@@ -21,4 +21,8 @@
 
 # ─────────────────────────────────────────────────────────────────────────────
 # Type coercion in AVG — pre-existing limitation (xsd:double cast of string literals)
+# Reason: AVG() over IF(isNumeric(?x), xsd:double(?x), 0) fails because the
+#   SPARQL-to-SQL translator emits a plain CASE expression without explicit
+#   CAST(... AS float8). Fix tracked in #SPARQL-AVG-CAST.
+# Issue: https://github.com/grove/pg-ripple/issues/460
 http://www.w3.org/2009/sparql/docs/tests/data-sparql11/aggregates/manifest#agg-err-02  AVG(IF(isNumeric,xsd:double cast)) — string-to-double cast in aggregation not yet supported
